requirements.txt
python-dotenv>=1.0.0
pytz>=2025.2
asyncio==4.0.0
requests==2.32.5
tenacity>=9.1.2
flask>=2.0.0
pandas>=1.3.0
numpy>=1.21.0
bpx-py==2.0.11
Flask-Cors>=5.0.0
redis>=5.0.0

# matplotlib>=3.5.0  # 不再需要，使用前端Chart.js绘图

# Forked edgeX official python sdk with post_only limit order
git+https://github.com/your-quantguy/edgex-python-sdk.git@7ce85b6ebfa9df0a6f6252d93e10d106927257a9#egg=edgex-python-sdk

# Lighter exchange SDK
git+https://github.com/elliottech/lighter-python.git@d0009799970aad54ebb940aa3dc90cbc00028c54